Add a Docker file and 'docker_builder.sh' script
One of the project targets is to create a code that runs fast.
To do that, we will need the latest version of the gcc compiler.
A Docker container is a good solution for that end. It allows compiling
the code on a container with the latest version of the operation system
(Ubuntu in our case), and install there whatever tools we want for
development.
Another advantage is that it allows running the code on every OS,
without taking care to dependency issues.
This ability might be added on a follow-up commit.

Add a Dockerfile that builds an image of the latest Ubuntu version
(22.04-jammy), and install there the latest version of gcc (11).

Add a 'docker_builder.sh' that creates a docker image if it does not
exist and then compiles the code in the container by running the
'build.sh' added in the previous commit.
Note that the container working directory is bound to the project
directory, so all the compilation dump files will be visible on the
'build/' folder of the host.

Signed-off-by: Moran Shoeg <morantr@gmail.com>