Reorganize the project files
Change the organization of the project files to a tree in a form of:
- project folder
    - build
    - tools
    - src
        - external
            - include
            - external cpp files
          main.cpp

This form is commonly used and is more convenient for scaling up the
project.

Note that now, the base64 and the sha256 are linked as a library.
The binary files moved to "build/bin/" and the libraries moved to
"build/lib/".

Signed-off-by: Moran Shoeg <morantr@gmail.com>