Add CMake C and CPP compilation flags
Add very restrictive compilation flags that allow the detection of many issues with the code.

Add the "Wfatal-errors" to fail the compilation on the first compilation error since it is easier
to solve compilation issues where there is a single error printed on the screen and not tons
of them.

Signed-off-by: Moran Shoeg <morantr@gmail.com>