Add the BaseOperationsUtils utility class
Add a static utility class called BaseOperationsUtils.
This class contains functions that will help in the calculations of another custom base arithmetic
operation with the restriction of the project, such as:

- Do the calculations on strings since the permutations could reach values that exceed unsigned
integer limits.

- The functions need to be fast as possible.

The added functions are:
1. decimal_to_base_x() - convert a decimal to a new custom base. This function is needed to let some
worker threads or clients do limited calculations from value x to value y and coordinate the
calculations between them.

2. sum_base_x_integers() - summarize two strings containing numbers in a custom base.
This function is also needed for the same reason as the previous function.

3. increment_base_x_integer() - increment a string containing an integer of a custom base by 1.
This function is needed for iterating on all permutations.

Add unit tests for all functions in the new class.

Signed-off-by: Moran Shoeg <morantr@gmail.com>