Add the HashGenerator class
The class is responsible for generating hashes of every string permutation.
It does it by adding the salt and pepper strings, performing SHA-256 encryption, and encoding the
result to base64.
To use the class need to call set_initial_permutation(str) to initialize the permutation of which
the class will start creating hashes from.
To generate the next permutation hash, need to call get_next_permutation_hash().

Signed-off-by: Moran Shoeg <morantr@gmail.com>