Add UiUtils class with build_hash_rate_string() function in it
The 'UiUtils' class purpose is to have any UI-related functions.
Currently, added only the build_hash_rate_string. It takes two integers that were set in a time difference 'update_rate' in seconds and calculates the hash rate.
This function will be used in a future commit.

Add also unit test to that function.

Signed-off-by: Moran Shoeg <morantr@gmail.com>