Optimizing the HashGenerator
After encrypting the permuted password, the HashGenerator encodes it
with base64. Upon further consideration, it appears that this step is
unnecessary. Instead, the hashed password list can be decoded from
base64 once, and the encryption result can be compared directly to the
decoded hash. This operation alone improves performance by 6X.

To achieve this, remove the base64 encoding step from the HashGenerator
and return the SHA-256 output directly.

Additionally, add a move operation to the move constructor
initialization, which was previously missing.

Signed-off-by: Moran Shoeg <morantr@gmail.com>