Adding the PollingScheduler Class
This class allows you to schedule a task and trigger a function when
a timeout occurs. To use this class, need to call the poll() function
frequently. This is acceptable because in the vast majority of cases,
the branch prediction mechanism will make the correct decision, and
anyway the thread needs to remain active for computation.

Signed-off-by: Moran Shoeg <morantr@gmail.com>