Add infrastructure for sending and receiving messages between threads
The infrastructure is based on a mutex mechanism and should be faster
than any other solution. The infrastructure implements separate
thread-safe and non-thread-safe functions for sending and receiving,
instead of general functions, to prevent internal branching, and by
that, increase performance.

The non-thread-safe version does not use a mutex for sending and
receiving messages, achieving better performance. This is useful when
the main thread is running a general code and needs to communicate with
it in a general way.

Signed-off-by: Moran Shoeg <morantr@gmail.com>