Add the HashCrackerThread class.
This class is responsible for computing the generation of hashes and
comparing them to the known hash list. If a match is found, the password
is cracked, and the object reports it to the management layer above it.

This class's work can be run on a thread context or in the context of
the object creator for efficient use of computing power.

Signed-off-by: Moran Shoeg <morantr@gmail.com>