Changes and refactors UiUtils::build_hash_rate_string().
The next commit will change the main function to use the
HashCrackerThread instead of the HashGenerator.
In the new implementation, the rate of hashes per second will be
provided instead of counters. This allows simplifying the function.

In addition, extra optimization was done by removing the power
calculation and using a lookup table for everything, which improves
performance.

Update the unit test of that function accordingly.

Signed-off-by: Moran Shoeg <morantr@gmail.com>