Adapt code to run on Windows platform
This commit includes changes to make the code compatible with the Windows operating system.
These changes include:

1. Adding a new configuration in the .vscode/c_cpp_properties.json file for the MinGW compiler.
2. Replacing the "uint" data type with "uint32_t" in the method signatures of BaseOperationsUtils
.h and .cpp files, unit.cpp. This ensures compatibility with Windows, where compilation failed due
to the use of "uint" (which worked fine on Linux).
3. Adding a portable_endian.h header file for endian conversion functions and including it in the
sha256 library. The library used "endian.h" which exists only on Linux. This open-source file
allows the use of all definitions from "endian.h" file in other operating systems.

Signed-off-by: Moran Shoeg <morantr@gmail.com>