Fix hanging thread issue when all tasks are finished
Currently, the main loop only checks if the main thread has finished its
task. This can cause problems if the main loop ends before all threads
have finished, as an unfinished thread may not be able to handle
messages such as "FINISHED_TASK". This can cause the process to hang
without a way to release it.

To fix this, we should remove the main loop's check for the main
thread's completion and instead break the thread if the sum of all
thread hash rates is zero. When we receive the "FINISHED_TASK" message
event from a thread, we should change its hash rate to zero, since the
thread is stopped and we will not receive any more hash rate updates
from it. This will ensure that the thread's last reported hash rate is
retained.

We can also remove the unused HashCrackerManager::m_finished_task and
its getter function, as they are not needed and do not contribute to
the code's functionality.

Signed-off-by: Moran Shoeg <morantr@gmail.com>