Fix a issue where REMOVE_HASH_FROM_LIST causes a fatal error message
When HashCrackerThread receives a REMOVE_HASH_FROM_LIST message, it will
fail to find the hash in its internal list. This is because the message
is sent back to the thread that originally sent the HASH_DISCOVERY
message, which has already removed the hash.

The root cause of this issue is a wrong comparison in the
HashCrackerManager::init() function, where the function compares the
id to '0' instead of the thread's internal id. To fix the issue, the
comparison has been corrected to use the internal id.

Additionally, logs aid with debugging the issue have been added.

Signed-off-by: Moran Shoeg <morantr@gmail.com>